@@ -5,10 +5,10 @@ const nextConfig = {
     remotePatterns: [
       {
         protocol: "https",
-        hostname: "**",
+        hostname: "images.ctfassets.net",
       },
     ],
   },
 };
 
-module.exports = nextConfig;
+export default nextConfig;
