@@ -1,5 +1,5 @@
 import type { ReactNode } from "react";
-
+import { UserDetailsProvider } from "@/components/contexts/UserDetailsContext";
 import Footer from "@/components/Footer";
 import Header from "@/components/Header";
 import PagePlaceholder from "@/components/PagePlaceholder";
